Moving input file 1 level above, improvements
Removing TUV input file as it is now 1 level above outside the repo.
This is to be able to incorporate numTUVrxn as submodule in the INPUTS
folder of TUV 5.2.
Adding check for slash as indicator for folder path in ifile string. If
no path is present '../' will be added after the programme argument is
read. 
Removing query for switch argument when left empty. This way an obsolete
2nd programme argument will automatically lead to the standard case,
where all reaction switches are left unchanged in the TUV input file.
Updating Readme.