Updated file handling in Julia script
Adding default folder as parent folder, if numTUVrxn.jl is called from
the repo folder. Using routines from module fhandle in repo auxdata to
check existence of file and read the content.
New option to give a path of a md file to the second script argument. In
this case, the TUV input file reaction numbers are compared to the
reaction numbers in the md file and warning will be issued for
mismatches.
Updated TUV input file and new md sample file.